Refactoring IT and naming changes
